いろいろ(未遂) タイトル決定 タイトル画面をよさげに デフォルトファイル埋め込みを抹殺 htmlを”呼べなかった”

+++ SchoolFestival2017/GameScene.hpp
@@ -17,7 +17,7 @@ inline double vec2ToRadian(const Vec2& v){
 class GameScene:public MySceneManager::Scene{
 private:
 	const int hitsize	=50;
-	const int seedslimit=5;
+	const size_t seedslimit=5;
 public:
 	Drag drag;
 	Unit unit;
